requirements.txt
pygame==1.9.4
PyYAML==3.13